Fix critical bug in handling hashes and add more auditing Original rev  : rc@rc0.org.uk--2004-mairix/mairix--dev--0.15--patch-12
This patch fixes a critical bug, where the hash key read from the file database
wasn't propagated to the in-core one.  Thus any tokens added during a new
indexing run were added to the token tables using a different hash function to
that used before.  This led to duplicate tokens, which eventually caused
strange crashes and meant that searches were bogus because they would only see
some of the potential matches.

Additional auditing is added during the integrity checks to increase the
likelihood of catching this kind of problem in the future.