[PATCH] mairix --version patch
G'day,

Testing 0.17-pre1 now ... after building it into a Debian package using
the debian/ directory from 0.16, and installing it, I wanted to verify
the version.  xterm was 24x80 and --help scrolled off.

So I propose this patch to add --version support; to stdout rather than
stderr, which seems to be the convention (checked hello and other
tools).

--
James Cameron
http://ftp.hp.com.au/sigs/jc/