cache uncompressed mbox data
Getting search results from compressed mbox files can be very slow
because mairix is decompressing the file over and over again to
copy each message out.  This patch creates a cache that decompresses
to a tmp file and returns an mmap of the tmp file to the caller.

By default only one file is cached, but changing this is a simple
define.  It seems that mairix is smart when pulling results out,
so only one cache file is needed.

Before this patch, a search that returned 389 messages from
various compressed archives took 1m28s.  With the patch it needs
only 3 seconds.