Implement -F/--fast-index flag
For maildir and MH folders, this flag avoids doing a stat() on each message.
This should give a noticeable improvement in indexing speed when the majority
of messages are unchanged and indexing is occurring very frequently.  However,
there is a risk that mairix will not notice changes to an existing message.
However, such an omission can easily be cured by a later indexing run without
this new flag.