Limit number of messages that are examined when an end boundary is missing.
When a large mbox contains a message with a broken attachment such as a
missing end boundary, mairix keeps looking through the rest of the mbox
in hope of finding the end boundary.  This often futile activity takes
a long time if the mbox is large.  The following patch restricts the
search to the next 100 messages in the mbox.