Fix segfault when last char is not a newline
I stumbled across a small bug when I tried out Mairix. It happens when an email
file does not have a newline as the last character, sinc you do strchr(string,
'\n'), which returns zero if it can't find '\n', resulting in a segfault.

I've attached a patch that fixes the issue for me.