Fix faultly mbox message separators
Jonathan writes:
I noticed an issue with mairix today when doing a search with output into an
mbox file.

Technically, the message separator for mbox files is “\n\nFrom “, i.e., there’s
always supposed to be a blank line before the “From “ line, but the mbox being
generated by mairix search is missing the blank line at the end of some
messages, with the result that some programs that parse mbox files were missing
some message breaks and glomming messages together.

The attached patch fixes this by ensuring that there’s a blank line at the end
of every message in the mbox.