Option to follow symlinks to mboxes in rc file
Add .mairixrc option "follow_mbox_symlinks" which allows the user to
restore the behaviour disabled by commit 74ebca5.

This is particularly useful for users of git-annex.  It is possible to
work around mairix's symlink-skipping behaviour [1], but a simple
configuration file option to restore the old behaviour is much more
practical.

This patch does not change the default behaviour of mairix for users who
do not explicitly include "follow_mbox_symlinks" in their .mairixrc.

[1] https://joeyh.name/blog/entry/moving_my_email_archives_and_packages_to_git-annex/