Don't warn when MH subfolder names look like message names.
Previously, any file in an MH folder with a numeric name was treated as
a message and if it was found to be a direcotry instead of a regular
file (as a subfolder with a numeric name would be), we complained about
it. Now, suppress these warnings.

Fixes rc0/mairix#24 and https://bugs.debian.org/700906