Raise the limit of 65536 mboxes and 65536 messages per mbox.
This is an incompatible change to the index format and requires it
to be rebuilt.

The mbox index and message index used to be packed into a single
"unsigned int" (presumed to be 32 bits) in a slot that is used either
as an offset to cdata (for maildir and others) or as this packed value
(for mbox). After this change, it is always an offset to data, and
in the case of mbox the data contain the mbox and message index number,
both as 32 bit values. I thought about making them 64 bits each for
good measure but I imagine that is not necessary.

Thanks to https://github.com/idallen for looking into this
issue at https://github.com/rc0/mairix/issues/23 and
https://github.com/vandry/mairix/issues/30 and testing changes
to the encoding format.